+++ main.tf
@@ -11,7 +11,7 @@ terraform {
 resource "aws_cloudformation_stack" "oe_patterns_pixelfed" {
   name = var.stack_name
 
-  template_url = "https://s3.amazonaws.com/awsmp-fulfillment-cf-templates-prod/93b3eaa6-e079-44f0-ab14-0b7dd0d3809f/5e473f9f83ef4c1998972f0af3a547fa.template"
+  template_url = "https://awsmp-cft-211125678794-1707910187780.s3.us-east-1.amazonaws.com/b0d1c344-27ad-4a88-b495-28e833322424/b0d1c344-27ad-4a88-b495-28e833322424/template.yaml"
 
   capabilities = ["CAPABILITY_NAMED_IAM"]
 
@@ -25,8 +25,11 @@ resource "aws_cloudformation_stack" "oe_patterns_pixelfed" {
     ActivityPubEnabled                 = var.activity_pub_enabled
     AlbCertificateArn                  = var.alb_certificate_arn
     AlbIngressCidr                     = var.alb_ingress_cidr
+    AsgDataVolumeBackupRetentionPeriod = var.asg_data_volume_backup_retention_period
+    AsgDataVolumeBackupVaultArn        = var.asg_data_volume_backup_vault_arn
     AsgDataVolumeSize                  = var.asg_data_volume_size
     AsgDataVolumeSnapshot              = var.asg_data_volume_snapshot
+    AsgDiskUsageAlarmThreshold         = var.asg_disk_usage_alarm_threshold
     AsgInstanceType                    = var.asg_instance_type
     AsgKeyName                         = var.asg_key_name
     AsgReprovisionString               = var.asg_reprovision_string